Move collections models to models.py
